requirements.txt
python-dotenv==1.0.0
discord==2.3.1
lyricsgenius==3.0.1